Allow skipping keys that don't have message documentation (fixes #54)
This would allow us to enable banana-checker on repositories that are
currently not fully documented, ensuring that we don't regress any
further when new messages are added, and can gradually come into proper
compliance.

A list of messages that should be skipped from checking can be specified
in the Gruntfile.js with the "skipIncompleteMessageDocumentation"
option.

Tests included.