build: Remove eslint file whitelist in Grunfile
Let eslint traverse the directory instead, the same way it would
when callling 'eslint .' or 'eslint --fix .' from e.g. CLI
or inside IDEs.

The only directory to exclude is node_modules, which is excluded
by default. If and when others need to be ignored, we'd add specify
them in an .eslintignore file.