build: Update packages and CI config (#65)
* build: Update eslint-config-wikimedia to 0.9.0

This release also provides eslint and plugins as dependencies,
so we can use those as-is going forward.

* build: Remove eslint file whitelist in Grunfile

Let eslint traverse the directory instead, the same way it would
when callling 'eslint .' or 'eslint --fix .' from e.g. CLI
or inside IDEs.

The only directory to exclude is node_modules, which is excluded
by default. If and when others need to be ignored, we'd add specify
them in an .eslintignore file.

* build: Remove deprecated 'sudo: false' setting from Travis CI

Travis CI has unified their two infrastructures (GCE/VM based, where
sudo is allowed) and the Docker container infra (where sudo is
not available).

This setting no longer has any effect.

https://blog.travis-ci.com/2018-11-19-required-linux-infrastructure-migration

Also, reorder the node_js versions so that current versions are
tested first for faster feedback on common problems.