Allow ignoring missing translations of blank source messages (#74)
For the requireCompleteTranslationLanguages and
requireCompleteTranslationMessages features, it doesn't make sense to
fail when a message whose source text is blank doesn't have a
translation. These messages aren't intended to be translated, and even
if they were translated as blanks, that would also fail because
disallowBlankTranslations is true by default.

This adds ignoreMissingBlankTranslations and sets it to true by default.
If enabled, a message is only considered missing if its source text is
non-empty. Blank source message do still require documentation.